Fix circular dependencies again
