+++ packages/react-native-sortables/src/debug/providers/DebugProvider.tsx
@@ -6,8 +6,14 @@ import {
   useRef,
   useState
 } from 'react';
+import type { View } from 'react-native';
 import type { SharedValue } from 'react-native-reanimated';
-import { cancelAnimation, makeMutable, runOnUI } from 'react-native-reanimated';
+import {
+  cancelAnimation,
+  makeMutable,
+  runOnUI,
+  useAnimatedRef
+} from 'react-native-reanimated';
 
 import { useDebouncedStableCallback } from '../../hooks';
 import { createProvider } from '../../providers/utils';
@@ -25,6 +31,7 @@ const { DebugProvider, useDebugContext } = createProvider('Debug', {
   const debugIdRef = useRef(0);
   const debugViewsRef = useRef<DebugViews>({});
   const observersRef = useRef(new Set<(views: DebugViews) => void>());
+  const debugOutletRef = useAnimatedRef<View>();
 
   const getNextKey = useCallback(() => debugIdRef.current++, []);
 
@@ -206,6 +213,7 @@ const { DebugProvider, useDebugContext } = createProvider('Debug', {
   return {
     enabled,
     value: {
+      debugOutletRef,
       useDebugCross,
       useDebugLine,
       useDebugLines: useDebugLines as DebugContextType['useDebugLines'],
