+++ packages/react-native-sortables/src/debug/providers/index.ts
@@ -1,5 +1,2 @@
-export {
-  DebugOutletProvider,
-  useDebugOutletContext
-} from './DebugOutletProvider';
+export { default as DebugOutlet } from './DebugOutlet';
 export * from './DebugProvider';
